fix(frontend): component moved
